pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "shantanu-jalkote",email = "shantanu.jalkote@techolution.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "google-cloud-aiplatform (>=1.91.0,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "motor (>=3.7.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.2"

